Moving example entities
